+++ pkg/client/cli/cmd_status.go
@@ -2,10 +2,13 @@ package cli
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/telepresenceio/telepresence/v2/pkg/client/scout"
+	"io"
 	"net"
-	"strings"
+	"time"
 
 	"github.com/spf13/cobra"
 	empty "google.golang.org/protobuf/types/known/emptypb"
@@ -16,117 +19,166 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/iputil"
 )
 
+type statusInfo struct {
+	json bool
+	out  io.Writer
+}
+
+type statusOutput struct {
+	DaemonStatus daemonStatus    `json:"root_daemon"`
+	UserDaemon   connectorStatus `json:"user_daemon"`
+}
+
+type daemonStatus struct {
+	Running           bool             `json:"running,omitempty"`
+	Version           string           `json:"version,omitempty"`
+	APIVersion        int32            `json:"api_version,omitempty"`
+	DNS               *daemonStatusDNS `json:"dns,omitempty"`
+	AlsoProxySubnets  []string         `json:"also_proxy_subnets,omitempty"`
+	NeverProxySubnets []string         `json:"never_proxy_subnets,omitempty"`
+}
+
+type daemonStatusDNS struct {
+	LocalIP         net.IP        `json:"local_ip,omitempty"`
+	RemoteIP        net.IP        `json:"remote_ip,omitempty"`
+	ExcludeSuffixes []string      `json:"exclude_suffixes,omitempty"`
+	IncludeSuffixes []string      `json:"include_suffixes,omitempty"`
+	LookupTimeout   time.Duration `json:"lookup_timeout_in_nanos,omitempty"`
+}
+
+type connectorStatus struct {
+	Running           bool                           `json:"running,omitempty"`
+	Version           string                         `json:"version,omitempty"`
+	APIVersion        int32                          `json:"api_version,omitempty"`
+	Executable        string                         `json:"executable,omitempty"`
+	InstallID         string                         `json:"install_id,omitempty"`
+	AmbassadorCloud   connectorStatusAmbassadorCloud `json:"ambassador_cloud"`
+	Status            string                         `json:"status,omitempty"`
+	Error             string                         `json:"error,omitempty"`
+	KubernetesServer  string                         `json:"kubernetes_server,omitempty"`
+	KubernetesContext string                         `json:"kubernetes_context,omitempty"`
+	Intercepts        []connectStatusIntercept       `json:"intercepts,omitempty"`
+}
+
+type connectorStatusAmbassadorCloud struct {
+	Status    string `json:"status,omitempty"`
+	UserID    string `json:"user_id,omitempty"`
+	AccountID string `json:"account_id,omitempty"`
+}
+
+type connectStatusIntercept struct {
+	Name   string `json:"name,omitempty"`
+	Client string `json:"client,omitempty"`
+}
+
 func statusCommand() *cobra.Command {
-	return &cobra.Command{
+	s := &statusInfo{}
+	cmd := &cobra.Command{
 		Use:  "status",
 		Args: cobra.NoArgs,
 
 		Short: "Show connectivity status",
-		RunE:  status,
+		RunE:  s.status,
 	}
+	flags := cmd.Flags()
+	flags.BoolVarP(&s.json, "json", "j", false, "output as json object")
+	return cmd
 }
 
 // status will retrieve connectivity status from the daemon and print it on stdout.
-func status(cmd *cobra.Command, _ []string) error {
-	if err := daemonStatus(cmd); err != nil {
+func (s *statusInfo) status(cmd *cobra.Command, _ []string) error {
+	s.out = cmd.OutOrStdout()
+	ctx := cmd.Context()
+
+	ds, err := s.daemonStatus(ctx)
+	if err != nil {
 		return err
 	}
 
-	if err := connectorStatus(cmd); err != nil {
+	cs, err := s.connectorStatus(ctx)
+	if err != nil {
 		return err
 	}
 
+	if s.json {
+		return s.printJSON(ds, cs)
+	}
+	s.printText(ds, cs)
 	return nil
 }
 
-func daemonStatus(cmd *cobra.Command) error {
-	out := cmd.OutOrStdout()
-
-	err := cliutil.WithStartedNetwork(cmd.Context(), func(ctx context.Context, daemonClient daemon.DaemonClient) error {
+func (s *statusInfo) daemonStatus(ctx context.Context) (daemonStatus, error) {
+	ds := daemonStatus{}
+	err := cliutil.WithStartedNetwork(ctx, func(ctx context.Context, daemonClient daemon.DaemonClient) error {
+		ds.Running = true
 		var err error
-		status, err := daemonClient.Status(cmd.Context(), &empty.Empty{})
+		status, err := daemonClient.Status(ctx, &empty.Empty{})
 		if err != nil {
 			return err
 		}
-		version, err := daemonClient.Version(cmd.Context(), &empty.Empty{})
+		version, err := daemonClient.Version(ctx, &empty.Empty{})
 		if err != nil {
 			return err
 		}
 
-		fmt.Fprintln(out, "Root Daemon: Running")
-		fmt.Fprintf(out, "  Version   : %s (api %d)\n", version.Version, version.ApiVersion)
+		ds.Running = true
+		ds.Version = version.Version
+		ds.APIVersion = version.ApiVersion
 		if obc := status.OutboundConfig; obc != nil {
+			ds.DNS = &daemonStatusDNS{}
 			dns := obc.Dns
-			fmt.Fprintf(out, "  DNS       :\n")
 			if dns.LocalIp != nil {
 				// Local IP is only set when the overriding resolver is used
-				fmt.Fprintf(out, "    Local IP        : %v\n", net.IP(dns.LocalIp))
+				ds.DNS.LocalIP = dns.LocalIp
 			}
-			fmt.Fprintf(out, "    Remote IP       : %v\n", net.IP(dns.RemoteIp))
-			fmt.Fprintf(out, "    Exclude suffixes: %v\n", dns.ExcludeSuffixes)
-			fmt.Fprintf(out, "    Include suffixes: %v\n", dns.IncludeSuffixes)
-			fmt.Fprintf(out, "    Timeout         : %v\n", dns.LookupTimeout.AsDuration())
-			fmt.Fprintf(out, "  Also Proxy : (%d subnets)\n", len(obc.AlsoProxySubnets))
+			ds.DNS.RemoteIP = dns.RemoteIp
+			ds.DNS.ExcludeSuffixes = dns.ExcludeSuffixes
+			ds.DNS.IncludeSuffixes = dns.IncludeSuffixes
+			ds.DNS.LookupTimeout = dns.LookupTimeout.AsDuration()
 			for _, subnet := range obc.AlsoProxySubnets {
-				fmt.Fprintf(out, "    - %s\n", iputil.IPNetFromRPC(subnet))
+				ds.AlsoProxySubnets = append(ds.AlsoProxySubnets, fmt.Sprintf("%v", iputil.IPNetFromRPC(subnet)))
 			}
-			fmt.Fprintf(out, "  Never Proxy: (%d subnets)\n", len(obc.NeverProxySubnets))
 			for _, subnet := range obc.NeverProxySubnets {
-				fmt.Fprintf(out, "    - %s\n", iputil.IPNetFromRPC(subnet))
+				ds.NeverProxySubnets = append(ds.NeverProxySubnets, fmt.Sprintf("%v", iputil.IPNetFromRPC(subnet)))
 			}
 		}
 		return nil
 	})
 	if err != nil {
 		if errors.Is(err, cliutil.ErrNoNetwork) {
-			fmt.Fprintln(out, "Root Daemon: Not running")
-			return nil
+			ds.Running = false
+			return ds, nil
 		}
-		return err
+		return ds, err
 	}
-	return nil
+	return ds, nil
 }
 
-func connectorStatus(cmd *cobra.Command) error {
-	out := cmd.OutOrStdout()
-
-	err := cliutil.WithStartedConnector(cmd.Context(), false, func(ctx context.Context, connectorClient connector.ConnectorClient) error {
-		fmt.Fprintln(out, "User Daemon: Running")
-
-		type kv struct {
-			Key   string
-			Value string
-		}
-		var fields []kv
-		defer func() {
-			klen := 0
-			for _, kv := range fields {
-				if len(kv.Key) > klen {
-					klen = len(kv.Key)
-				}
-			}
-			for _, kv := range fields {
-				vlines := strings.Split(strings.TrimSpace(kv.Value), "\n")
-				fmt.Fprintf(out, "  %-*s: %s\n", klen, kv.Key, vlines[0])
-				for _, vline := range vlines[1:] {
-					fmt.Fprintf(out, "    %s\n", vline)
-				}
-			}
-		}()
-
+func (s *statusInfo) connectorStatus(ctx context.Context) (connectorStatus, error) {
+	cs := connectorStatus{}
+	err := cliutil.WithStartedConnector(ctx, false, func(ctx context.Context, connectorClient connector.ConnectorClient) error {
+		cs.Running = true
 		version, err := connectorClient.Version(ctx, &empty.Empty{})
 		if err != nil {
 			return err
 		}
-		fields = append(fields, kv{"Version", fmt.Sprintf("%s (api %d)", version.Version, version.ApiVersion)})
-		fields = append(fields, kv{"Executable", version.Executable})
+		cs.Version = version.Version
+		cs.APIVersion = version.ApiVersion
+		cs.Executable = version.Executable
+		reporter := scout.NewReporter(ctx, "cli")
+		cs.InstallID = reporter.InstallID()
 
 		if !cliutil.HasLoggedIn(ctx) {
-			fields = append(fields, kv{"Ambassador Cloud", "Logged out"})
-		} else if _, err := cliutil.GetCloudUserInfo(ctx, false, true); err != nil {
-			fields = append(fields, kv{"Ambassador Cloud", "Login expired (or otherwise no-longer-operational)"})
+			cs.AmbassadorCloud.Status = "Logged out"
 		} else {
-			fields = append(fields, kv{"Ambassador Cloud", "Logged in"})
+			userInfo, err := cliutil.GetCloudUserInfo(ctx, false, true)
+			if err != nil {
+				cs.AmbassadorCloud.Status = "Login expired (or otherwise no-longer-operational)"
+			} else {
+				cs.AmbassadorCloud.Status = "Logged in"
+				cs.AmbassadorCloud.UserID = userInfo.Id
+				cs.AmbassadorCloud.AccountID = userInfo.AccountId
+			}
 		}
 
 		status, err := connectorClient.Status(ctx, &empty.Empty{})
@@ -135,37 +187,112 @@ func connectorStatus(cmd *cobra.Command) error {
 		}
 		switch status.Error {
 		case connector.ConnectInfo_UNSPECIFIED, connector.ConnectInfo_ALREADY_CONNECTED:
-			fields = append(fields, kv{"Status", "Connected"})
+			cs.Status = "Connected"
 		case connector.ConnectInfo_MUST_RESTART:
-			fields = append(fields, kv{"Status", "Connected, but must restart"})
+			cs.Status = "Connected, but must restart"
 		case connector.ConnectInfo_DISCONNECTED:
-			fields = append(fields, kv{"Status", "Not connected"})
+			cs.Status = "Not connected"
 			return nil
 		case connector.ConnectInfo_CLUSTER_FAILED:
-			fields = append(fields, kv{"Status", "Not connected, error talking to cluster"})
-			fields = append(fields, kv{"Error", status.ErrorText})
+			cs.Status = "Not connected, error talking to cluster"
+			cs.Error = status.ErrorText
 			return nil
 		case connector.ConnectInfo_TRAFFIC_MANAGER_FAILED:
-			fields = append(fields, kv{"Status", "Not connected, error talking to in-cluster Telepresence traffic-manager"})
-			fields = append(fields, kv{"Error", status.ErrorText})
+			cs.Status = "Not connected, error talking to in-cluster Telepresence traffic-manager"
+			cs.Error = status.ErrorText
 			return nil
 		}
-		fields = append(fields, kv{"Kubernetes server", status.ClusterServer})
-		fields = append(fields, kv{"Kubernetes context", status.ClusterContext})
-		intercepts := fmt.Sprintf("%d total\n", len(status.GetIntercepts().GetIntercepts()))
+		cs.KubernetesServer = status.ClusterServer
+		cs.KubernetesContext = status.ClusterContext
 		for _, icept := range status.GetIntercepts().GetIntercepts() {
-			intercepts += fmt.Sprintf("%s: %s\n", icept.Spec.Name, icept.Spec.Client)
+			cs.Intercepts = append(cs.Intercepts, connectStatusIntercept{
+				Name:   icept.Spec.Name,
+				Client: icept.Spec.Client,
+			})
 		}
-		fields = append(fields, kv{"Intercepts", intercepts})
-
 		return nil
 	})
 	if err != nil {
 		if errors.Is(err, cliutil.ErrNoUserDaemon) {
-			fmt.Fprintln(out, "User Daemon: Not running")
-			return nil
+			cs.Running = false
+			return cs, nil
 		}
+		return cs, err
+	}
+	return cs, nil
+}
+
+func (s *statusInfo) printJSON(ds daemonStatus, cs connectorStatus) error {
+	output, err := json.Marshal(statusOutput{
+		DaemonStatus: ds,
+		UserDaemon:   cs,
+	})
+	if err != nil {
 		return err
 	}
+	s.println(string(output))
 	return nil
+}
+
+func (s *statusInfo) printText(ds daemonStatus, cs connectorStatus) {
+	s.printDaemonText(ds)
+	s.printConnectorText(cs)
+}
+
+func (s *statusInfo) printDaemonText(ds daemonStatus) {
+	if ds.Running {
+		s.println("Root Daemon: Running")
+		s.printf("  Version   : %s (api %d)\n", ds.Version, ds.APIVersion)
+		if ds.DNS != nil {
+			s.printf("  DNS       :\n")
+			if len(ds.DNS.LocalIP) > 0 {
+				s.printf("    Local IP        : %v\n", ds.DNS.LocalIP)
+			}
+			s.printf("    Remote IP       : %v\n", ds.DNS.RemoteIP)
+			s.printf("    Exclude suffixes: %v\n", ds.DNS.ExcludeSuffixes)
+			s.printf("    Include suffixes: %v\n", ds.DNS.IncludeSuffixes)
+			s.printf("    Timeout         : %v\n", ds.DNS.LookupTimeout)
+			s.printf("  Also Proxy : (%d subnets)\n", len(ds.AlsoProxySubnets))
+			for _, subnet := range ds.AlsoProxySubnets {
+				s.printf("    - %s\n", subnet)
+			}
+			s.printf("  Never Proxy: (%d subnets)\n", len(ds.NeverProxySubnets))
+			for _, subnet := range ds.NeverProxySubnets {
+				s.printf("    - %s\n", subnet)
+			}
+		}
+	} else {
+		s.println("Root Daemon: Not running")
+	}
+}
+
+func (s *statusInfo) printConnectorText(cs connectorStatus) {
+	if cs.Running {
+		s.println("User Daemon: Running")
+		s.printf("  Version         : %s (api %d)\n", cs.Version, cs.APIVersion)
+		s.printf("  Executable      : %s\n", cs.Executable)
+		s.printf("  Install ID      : %s\n", cs.InstallID)
+		s.printf("  Ambassador Cloud:\n")
+		s.printf("    Status    : %s\n", cs.AmbassadorCloud.Status)
+		s.printf("    User ID   : %s\n", cs.AmbassadorCloud.UserID)
+		s.printf("    Account ID: %s\n", cs.AmbassadorCloud.AccountID)
+		s.printf("  Status            : %s\n", cs.Status)
+		if cs.Error != "" {
+			s.printf("  Error             : %s\n", cs.Error)
+		}
+		s.printf("  Kubernetes server : %s\n", cs.KubernetesServer)
+		s.printf("  Kubernetes context: %s\n", cs.KubernetesContext)
+		s.printf("  Intercepts        : %d total\n", len(cs.Intercepts))
+		for _, intercept := range cs.Intercepts {
+			s.printf("    %s: %s\n", intercept.Name, intercept.Client)
+		}
+	}
+}
+
+func (s *statusInfo) printf(format string, a ...interface{}) {
+	_, _ = fmt.Fprintf(s.out, format, a...)
+}
+
+func (s *statusInfo) println(a ...interface{}) {
+	_, _ = fmt.Fprintln(s.out, a...)
 }